Create README.md
Provided above are three examples of MySQL exercises completed in the Udemy course "The Ultimate MySQL Bootcamp", which included completion of 10 other exercises similar to these. In each of the above exercises, tables are created and data definition as well as manipulation languages are run with a description of the objective task commented out above each set of code.